e2e-libs: fold the runtime tier into a single pass
snapshot.mjs, artifacts.mjs and karma-bundles.mjs each walked the same
(library x method x phase) space and each rebuilt it: 48 rollup builds for 21
distinct cells. They also each gated on a build of their own, so the set being
snapshotted was only conventionally the set inside the bundle that shipped —
build.mjs carried a comment asking future readers to keep the two configurations
in step, which is exactly the kind of promise that quietly stops being true.

Replace all three with runtime.mjs: one build per cell, handed to every consumer.

  gates      ES5 parse, core-js payload, no externals, non-empty injected set
  snapshot   the injected SET vs snapshots/<lib>.<method>.<phase>.txt
  pre-flight the bundle run in a fresh node process, all self-checks passing
  artifact   bundle.js + self-checking index.html + manifest row (raw/min/gz)
  karma      the same bundle + QUnit driver, in real IE11 where one exists

Now the set that is snapshotted, the bytes that are measured and the bundle that
runs in IE11 are the same build by construction. Coverage grows on the way:
pre-flight and artifacts now cover all 21 cells instead of the 9 artifacts.mjs
reached, so `pre` and `pre+post` get a node-executes gate and an uploadable page
they never had.

runtimeBuild returns `{ code, chunk, injected }` — the set, not a count. The
field is renamed because `if (!injections)` would have silently passed on an
empty array; a rename makes every stale caller fail loudly instead.

Timings stay out of this pass on purpose: minification, pre-flight child
processes and file writes land between consecutive builds and move the CPU state
each one starts from, which is precisely what makes cross-cell timings
comparable or not. pipeline.mjs keeps measuring in its own quiet process, and
keeps rebuilding its [C] stage — a deliberate cost, now documented as such.
throughput is untouched: different pipeline (no Babel), seven bundlers.

Verified: all 18 snapshots still match baselines authored by the previous
capture, which is the proof that the build configuration survived the move.
