+++ package.json
@@ -43,12 +43,10 @@
     "compat-node": "node tests/compat/node-runner",
     "compat-rhino": "npm run zxi tests/compat/rhino-adapter.mjs",
     "debug-get-dependencies": "npm run zxi time tests/debug-get-dependencies/debug-get-dependencies.mjs",
-    "e2e-libs": "run-s e2e-libs-check-exercise e2e-libs-snapshot e2e-libs-artifacts",
-    "e2e-libs-artifacts": "npm run zxi time cd tests/e2e-libs/artifacts.mjs --",
+    "e2e-libs": "run-s e2e-libs-check-exercise e2e-libs-runtime",
     "e2e-libs-check-exercise": "npm run zxi time cd tests/e2e-libs/check-exercise.mjs --",
-    "e2e-libs-karma-bundles": "npm run zxi time cd tests/e2e-libs/karma-bundles.mjs --",
     "e2e-libs-pipeline": "npm run zxi time cd tests/e2e-libs/pipeline.mjs --",
-    "e2e-libs-snapshot": "npm run zxi time cd tests/e2e-libs/snapshot.mjs --",
+    "e2e-libs-runtime": "npm run zxi time cd tests/e2e-libs/runtime.mjs --",
     "e2e-libs-throughput": "npm run zxi time cd tests/e2e-libs/throughput.mjs --",
     "lint": "run-s prepare lint-raw",
     "lint-raw": "run-s build-types test-eslint bundle-package test-publint",
