+++ tests/e2e-libs/karma.conf.cjs
@@ -1,14 +1,14 @@
 'use strict';
 // Karma config for the e2e-libs IE11 leg. Mirrors tests/unit-karma/karma.conf.js but trimmed to the
 // only target that matters here: real IE11. No Chromium/Firefox/WebKit (and so no Playwright dep) —
-// modern engines are already covered by artifacts.mjs's node pre-flight. This leg runs the full runtime
+// modern engines are already covered by runtime.mjs's node pre-flight. This leg runs the full runtime
 // matrix (every library × method × unplugin phase) on real IE11, where a usage-pure detection miss
-// cannot hide; karma-bundles.mjs invokes it once per bundle (one page per cell, so nothing is
+// cannot hide; runtime.mjs invokes it once per bundle (one page per cell, so nothing is
 // co-loaded) — see README.
 //
 // QUnit is the same karma-qunit@4 / qunit@2 stack the unit-karma job already runs green in IE11.
 // The bundles come in via `-f=` (absolute paths, comma-separated): each is a self-contained UMD from
-// runtimeBuild with a QUnit driver appended (see karma-bundles.mjs / harness.mjs).
+// runtimeBuild with a QUnit driver appended (see runtime.mjs / harness.mjs).
 const { sync: which } = require('which');
 
 const customLaunchers = {
@@ -20,7 +20,7 @@ const customLaunchers = {
 };
 
 // IE only, and only where it exists: on the windows CI runner (CI set) or a dev box with iexplore.
-// karma-bundles.mjs makes the same check before starting Karma, so this list is never empty here.
+// runtime.mjs makes the same check before starting Karma, so this list is never empty here.
 const browsers = [];
 if (process.env.CI || which('iexplore.exe', { nothrow: true })) browsers.push('IE_NFM');
